Feature/cjm 245 add job to post

+++ ConJob.Domain/Services/PostService.cs
@@ -7,6 +7,7 @@ using ConJob.Domain.Response;
 using ConJob.Domain.Services.Interfaces;
 using ConJob.Entities;
 using LinqKit;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using static ConJob.Domain.Response.EServiceResponseTypes;
 
@@ -17,17 +18,19 @@ namespace ConJob.Domain.Services
         private readonly IPostRepository _postRepository;
         private readonly IUserRepository _userRepository;
         private readonly ILikeRepository _likeRepository;
+        private readonly IJobRepository _jobRepository;
         private readonly IMapper _mapper;
         private readonly IFilterHelper<PostDetailsDTO> _filterHelper;
         private readonly IFilterHelper<PostDTO> _filterHelper2;
 
         public static int PAGE_SIZE { get; set; } = 1;
 
-        public PostService(IPostRepository postRepository, IUserRepository userRepository, ILikeRepository likeRepository, IMapper mapper, IFilterHelper<PostDetailsDTO> filterHelper, IFilterHelper<PostDTO> filterHelper2) 
+        public PostService(IPostRepository postRepository, IUserRepository userRepository, ILikeRepository likeRepository, IJobRepository jobRepository , IMapper mapper, IFilterHelper<PostDetailsDTO> filterHelper, IFilterHelper<PostDTO> filterHelper2) 
         {
             _postRepository = postRepository;
             _userRepository = userRepository;
             _likeRepository = likeRepository;
+            _jobRepository = jobRepository;
             _mapper = mapper;
             _filterHelper = filterHelper;
             _filterHelper2 = filterHelper2;
@@ -42,11 +45,11 @@ namespace ConJob.Domain.Services
                 var post = _mapper.Map<PostModel>(newPost);
                 post = await _postRepository.AddPostAsync(user, post);
                 serviceResponse.Data = _mapper.Map<PostDTO>(post);
+                serviceResponse.ResponseType = EResponseType.Success;
                 serviceResponse.Message = "Add post successfully";
             } catch (InvalidOperationException)
             {
-                serviceResponse.ResponseType = EResponseType.CannotCreate;
-                serviceResponse.Message = "Owner (User) of post not found.";
+                throw new InvalidOperationException("Owner (User) of post not found.");
             }
             return serviceResponse;
         }
@@ -58,21 +61,15 @@ namespace ConJob.Domain.Services
             {
                 var posts = _mapper.ProjectTo<PostDTO>(_postRepository.GetPostNotDeleted())
                         .AsNoTracking();
-
                 #region paging
                 var result = await _filterHelper2.ApplyPaging(posts, pageNo, PAGE_SIZE);
                 #endregion
-                
-                if (result != null)
-                {
-                    serviceResponse.Data = result;
-                    serviceResponse.Message = "Get all posts successfully!";
-                }
-                else
-                {
-                    serviceResponse.ResponseType = EResponseType.NotFound;
-                    serviceResponse.Message = "Post not found.";
-                }
+                serviceResponse.ResponseType = EResponseType.Success;
+                serviceResponse.Data = result;
+            }
+            catch (InvalidOperationException)
+            {
+                throw new InvalidOperationException("Post and/or Owner (User) not found.");
             }
             catch { throw; }
             return serviceResponse;
@@ -83,15 +80,14 @@ namespace ConJob.Domain.Services
             var serviceResponse = new ServiceResponse<PostDetailsDTO>();
             try
             {
+                serviceResponse.ResponseType = EResponseType.Success;
                 serviceResponse.Data = await _mapper.ProjectTo<PostDetailsDTO>(_postRepository.GetUserPosts(userId))
                     .AsNoTracking()
                     .FirstAsync(c => c.id == id);
-                serviceResponse.Message = "Get post by id successfully!";
             }
             catch (InvalidOperationException)
             {
-                serviceResponse.ResponseType = EResponseType.NotFound;
-                serviceResponse.Message = "Post and/or Owner (User) not found.";
+                throw new InvalidOperationException("Post and/or Owner (User) not found.");
             }
             catch { throw; }
             return serviceResponse;
@@ -102,10 +98,10 @@ namespace ConJob.Domain.Services
             var serviceResponse = new ServiceResponse<PostDetailsDTO>();
             try
             {
+                serviceResponse.ResponseType = EResponseType.Success;
                 serviceResponse.Data = await _mapper.ProjectTo<PostDetailsDTO>(_postRepository.GetPosts())
                     .AsNoTracking()
                     .FirstAsync(c => c.id == id);
-                serviceResponse.Message = "Get post by id successfully!";
             }
             catch (InvalidOperationException)
             {
@@ -126,19 +122,13 @@ namespace ConJob.Domain.Services
                 if (post != null)
                 {
                     post = await _postRepository.UpdateAsync(id, postDTO);
-                    serviceResponse.Data = _mapper.Map<PostDetailsDTO>(post);
+                    serviceResponse.ResponseType = EResponseType.Success;
                     serviceResponse.Message = "Update post successfully!";
-                }
-                else
-                {
-                    serviceResponse.ResponseType = EResponseType.NotFound;
-                    serviceResponse.Message = "You doesn't permission to edit this field.";
                 }
             }
             catch (InvalidOperationException)
             {
-                serviceResponse.ResponseType = EResponseType.NotFound;
-                serviceResponse.Message = "Post and/or Owner (User) not found.";
+                throw new InvalidOperationException("Post and / or Owner(User) not found.");
             }
             catch { throw; }
             return serviceResponse;
@@ -149,32 +139,63 @@ namespace ConJob.Domain.Services
             var serviceResponse = new ServiceResponse<object>();
             try
             {
-                var post = _postRepository.GetById(id);
-                await _postRepository.DeleteAsync(post.id);
-                serviceResponse.Message = "Delete post successfully!";
+                var post = await _postRepository.GetUserPosts(userId)
+                    .FirstAsync(c => c.id == id);
+                if (post != null)
+                {
+                    post = _postRepository.GetById(id);
+                    await _postRepository.DeleteAsync(post.id);
+                    serviceResponse.ResponseType = EResponseType.Success;
+                    serviceResponse.Message = "Delete post successfully!";
+                }
             }
-            catch (DbUpdateConcurrencyException)
+            catch (InvalidOperationException)
             {
-                serviceResponse.ResponseType = EResponseType.NotFound;
-                serviceResponse.Message = CJConstant.SOMETHING_WENT_WRONG;
+                throw new InvalidOperationException("Post and / or Owner(User) not found.");
             }
             catch { throw; }
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<object>> ActiveAsync(int userId, int id)
+        public async Task<ServiceResponse<object>> DeleteAsync(int id)
+        {
+            var serviceResponse = new ServiceResponse<object>();
+            try
+            {
+                var post = _postRepository.GetById(id);
+                if (post != null)
+                {
+                    await _postRepository.DeleteAsync(post.id);
+                    serviceResponse.ResponseType = EResponseType.Success;
+                    serviceResponse.Message = "Delete post successfully!";
+                }
+                else
+                {
+                    serviceResponse.ResponseType = EResponseType.NotFound;
+                    serviceResponse.Message = "Post not found";
+                }
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                throw new DbUpdateConcurrencyException(CJConstant.SOMETHING_WENT_WRONG);
+            }
+            catch { throw; }
+            return serviceResponse;
+        }
+
+        public async Task<ServiceResponse<object>> ActiveAsync(int id)
         {
             var serviceResponse = new ServiceResponse<object>();
             try
             {
                 var post = _postRepository.GetById(id);
                 await _postRepository.ActiveAsync(post.id);
+                serviceResponse.ResponseType = EResponseType.Success;
                 serviceResponse.Message = "Post has been successfully approved.";
             }
             catch (DbUpdateConcurrencyException)
             {
-                serviceResponse.ResponseType = EResponseType.NotFound;
-                serviceResponse.Message = CJConstant.SOMETHING_WENT_WRONG;
+                throw new DbUpdateConcurrencyException(CJConstant.SOMETHING_WENT_WRONG);
             }
             catch { throw; }
             return serviceResponse;
@@ -202,16 +223,12 @@ namespace ConJob.Domain.Services
                 // apply sorting and paging
                 var sortedPosts = _filterHelper.ApplySorting(posts, filterParameters.OrderBy);
                 var pagedPosts = await _filterHelper.ApplyPaging(sortedPosts, filterParameters.Page, filterParameters.Limit);
-                if (pagedPosts?.Items?.Any() == true)
-                {
-                    serviceResponse.Data = pagedPosts;
-                    serviceResponse.Message = "Get and filter posts successfully!";
-                }
-                else
-                {
-                    serviceResponse.ResponseType = EResponseType.NotFound;
-                    serviceResponse.Message = "Post and/or Owner (User) not found.";
-                }
+                serviceResponse.ResponseType = EResponseType.Success;
+                serviceResponse.Data = pagedPosts;
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                throw new DbUpdateConcurrencyException(CJConstant.SOMETHING_WENT_WRONG);
             }
             catch { throw; }
             return serviceResponse;
@@ -220,19 +237,61 @@ namespace ConJob.Domain.Services
         public async Task<ServiceResponse<int>> UserLikePost(int userId, int postId)
         {
             var serviceResponse = new ServiceResponse<int>();
-            var toCheck = _likeRepository.getLikeByUserPost(userId, postId);
-            if (toCheck == null)
+            try
             {
-                await _likeRepository.AddAsync(new LikeModel { 
-                                        user = _userRepository.GetById(userId), 
-                                        post = _postRepository.GetById(postId) });
-                serviceResponse.Message = "Like post successfully";
+                var toCheck = _likeRepository.getLikeByUserPost(userId, postId);
+                if (toCheck == null)
+                {
+                    await _likeRepository.AddAsync(new LikeModel
+                    {
+                        user = _userRepository.GetById(userId),
+                        post = _postRepository.GetById(postId)
+                    });
+                    serviceResponse.ResponseType = EResponseType.Success;
+                    serviceResponse.Message = "Like post successfully";
+                }
             }
-            else
+            catch (DbUpdateConcurrencyException)
             {
-                serviceResponse.Message = "User already like the post.";
+                throw new DbUpdateConcurrencyException("User already like the post.");
             }
-            serviceResponse.Data = _postRepository.CountLikePost(postId);
+            catch { throw; }
+            return serviceResponse;
+        }
+
+        public async Task<ServiceResponse<object>> AddJobToPost(int userId, int jobId, int postId)
+        {
+            var serviceResponse = new ServiceResponse<object>(); 
+            try
+            {
+                var post = await _postRepository.GetUserPosts(userId)
+                            .FirstOrDefaultAsync(c => c.id == postId);
+                var job = await _jobRepository.GetUserJobs(userId)
+                            .FirstOrDefaultAsync(c => c.id == jobId);
+                if (job != null && post != null)
+                {
+                    var toCheck = _postRepository.GetById(postId);
+                    if (toCheck.job_id == null)
+                    {
+                        await _postRepository.addJobToPostAsync(jobId, toCheck);
+                        serviceResponse.ResponseType = EResponseType.Success;
+                        serviceResponse.Message = "Add job to post successfully.";
+                    }
+                    else
+                    {
+                        throw new DbUpdateException(serviceResponse.Message = "Post already exist job.");
+                    }
+                }
+                else
+                {
+                    throw new DbUpdateException("Post and/or Job and/or Owner (User) not found.");
+                }
+            }
+            catch (InvalidOperationException)
+            {
+                throw new BadHttpRequestException(CJConstant.SOMETHING_WENT_WRONG);
+            }
+            catch { throw; }
             return serviceResponse;
         }
     }
